chore: enforce dump order
This commit makes sure, the order of the json dumps is consistent with
the order django will output it with the `dumpdata` command. It also makes sure
the dumps don't use `CRLF`.

Additionally (and in order for this to work), we override the
`setup.cfg` in dev, so we can set project specific env vars.