feat: full accounting document export
This commit extends the accounting export to also include the full
document in the same manner as the application export.

For this lots of refactoring was necessary. The `ApplicationParser` has
been moved out of the case app and renamed to `DocumenParser`.

Attached files are now instantly added to the `PDFWriter`. This allows
for a more thorough validation of files and fixes a bug, where files we
weren't able to merge to the PDF where marked as included. This bug
also affected the application export.

Also generation of the PDF file has been moved into the
`caluma_document_parser`-module, as it depends on the merger in the
`DocumentParser`-instance.