chore(API): refactor template upload
Since the refactoring of the build, the documented commands don't work
anymore. Additionally the path to the templates changes now dynamically
with the python version used. This commit puts all the templates into
one global template directory. The command is refactored, so that it's
sufficient to provide the file name. The path to the template will then
be constructed in the command.