fix: do never link users with organisation identities
This is the first step to resolve this issue. This fix prevents
organisation Identities to be linked with keycloak users. This could
potentially lead to a situation, where a user can make a keycloak
account with their email address, but then be unable to login to mySAGW,
if the Email is already set on an organisation Identity, due to the unique
constraint on the email field.